ENGINES: MaterialManager cleanup prior to TextureManager cleanup
Materials must be cleaned up before textures are deleted, lest there
be references to textures in the materials afterwards. This can cause
problems at exit, as the materials will still attempt to dereference
the (now deallocated) texture, resulting in a segfault.

Oh, and the NWN menu background has a very quick & dirty hack for
the animated background. It's a work in progress.